Adjust controller namespace and extract plane raycast method
